attempting to make oh-my-zsh install more scriptable

+++ install.sh
@@ -8,19 +8,30 @@ for i in "$@"; do
     esac
 done
 
-if [[ -z "$SHH" ]]; then
-    exec 3>&1 4>&2 
-else
-    exec 3>&1 4>&2 &>/dev/null
-fi
+shut_up () {
+    if [[ -z "$SHH" ]]; then
+        echo "bein loud"
+        exec 3>&1 4>&2 
+    else
+        echo "shuttin up"
+        exec 3>&1 4>&2 &>/dev/null
+    fi
+}
 
+echo_always () {
+    echo "$@" >&3
+}
+
+we_have () {
+    type "$@" >/dev/null 2>&1
+}
 
 create_directory () {
 	/bin/mkdir -p "$@" 2>&4
 }
 
 setup_directories () {
-	echo "Setting up familiar directory structure..." >&3
+	echo_always "Setting up familiar directory structure..."
 	create_directory "$PROJECT_DIR"
 	create_directory "$DOCS_DIR"
 }
@@ -30,12 +41,16 @@ get_packages () {
 }
 
 update_system () {
-    echo "Updating system..." >&3
+    echo_always "Updating system..."
     sudo apt-get update && sudo apt-get upgrade 2>&4
 }
 
 add_configure_vim () {
-    echo "setting up ~the superior editor~ ..." >&3
+    if [ -f ~/.vimrc ]; then
+        echo_always "vim already installed, skipping"
+        return
+    fi
+    echo_always "setting up ~the superior editor~ ..."
 	get_packages vim
 	echo "syntax on" > ~/.vimrc
 	echo "set ts=4 sw=4 expandtab smarttab smartindent" >> ~/.vimrc
@@ -45,14 +60,26 @@ add_configure_vim () {
 
 add_configure_zsh () {
     if [ -f "$HOME/.zshrc" ]; then
-        echo "zsh already setup, skipping..." >&3
+        echo_always "zsh already setup, skipping..."
         return
     fi
-    echo "Installing and configuring zsh..." >&3
-    get_packages zsh # TODO: oh-my-zsh
+    echo_always "Installing and configuring zsh..."
+    if [ ! we_have zsh ]; then
+        get_packages zsh # TODO: oh-my-zsh
+    fi
     #/usr/bin/chsh -s $(which zsh) "$USER"
-    /bin/sh -c "$(/usr/bin/wget https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh -O - >&3 2>&4)"
+    OLD_SHH="$SHH"
+    unset SHH
+    shut_up
+    /bin/sh -c "$(/usr/bin/wget https://raw.githubusercontent.com/robbyrussell/oh-my-zsh/master/tools/install.sh -O -)"
+    if [[ ! -z "$OLD_SHH" ]]; then
+        SHH="$OLD_SHH"
+    fi
+    shut_up
     echo "ZSH_THEME=\"jtriley\"" >> ~/.zshrc
+# TODO: add cd function
+# TODO: add fsl alias
+
 }
 
 add_configure_git () {
@@ -67,7 +94,7 @@ add_configure_ssh () {
 }
 
 add_configure_python () {
-    if type pip3 >/dev/null; then
+    if we_have pip3; then
         echo "Python installed, skipping..."
         return
     fi
@@ -76,16 +103,20 @@ add_configure_python () {
 }
 
 add_configure_fuck () {
-    if type thefuck >/dev/null; then
-        echo "fuck installed, skipping..."
+    if we_have thefuck; then
+        echo_always "fuck installed, skipping..."
     fi
     sudo pip3 install thefuck
     echo "eval \$(thefuck --alias)" >> ~/.zshrc
 }
 
 add_configure_cdh () {
-    echo "Adding C&DH repo..." >&3
     CDH_REPO="$PROJECT_DIR/LAICE_CDH"
+    if [ -d "$CDH_REPO" ]; then
+        echo_always "C&DH repo already installed at: $CDH_REPO"
+        return
+    fi
+    echo_always "Adding C&DH repo..."
 	get_packages cmake
 	get_packages cmake-curses-gui
 	get_packages libssl-dev
@@ -101,6 +132,7 @@ if [[ -z "$DOCS_DIR" ]]; then
 	DOCS_DIR="$HOME/doc"
 fi
 
+shut_up
 update_system
 setup_directories
 add_configure_git
@@ -114,4 +146,3 @@ get_packages build-essential
 if [[ ! -z "$ADD_CDH" ]]; then
     add_configure_cdh
 fi
-
